service_definition.md
I've made small changes to a couple of paragraphs so that the wording around service definition matches this: https://dfe-service-manual.herokuapp.com/service-owner-guidance

I changed the wording over there too. I've linked from the service manual page to this page for now, as I don't want to get into the constituent parts of a service and their definitions just yet.